reset: starfive-jh7110: use platform_ioremap_iomem_byname.
The reset module is scattered in several domains, and each address
segment may be located in the module device management.

Using devm_platform_get_and_ioremap_resource->devm_ioremap_resource will
cause the address of this segment to be occupied by the reset driver,
and other modules cannot be used, so use ioremap that can be mapped
multiple times instead.

Signed-off-by: samin <samin.guo@starfivetech.com>